requirements.txt
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121
jupyter
ipykernel
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
kaggle
tqdm
grad-cam
streamlit
opencv-python